Add new team videos, refactor form components.
