--- src/components/form/form-input.tsx
@@ -1,45 +1,0 @@
-import { Control } from "react-hook-form";
-import {
-  FormField,
-  FormItem,
-  FormLabel,
-  FormControl,
-  FormDescription,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-
-interface FormInputProps {
-  name: string;
-  label: string;
-  placeholder: string;
-  description?: string;
-  control: Control<any>;
-}
-
-export const FormInput = ({
-  name,
-  label,
-  placeholder,
-  description,
-  control,
-}: FormInputProps) => (
-  <FormField
-    control={control}
-    name={name}
-    render={({ field }) => (
-      <FormItem>
-        <FormLabel>{label}</FormLabel>
-        <FormControl>
-          <Input
-            placeholder={placeholder}
-            {...field}
-            className="min-h-14 w-full rounded-2xl bg-black/[0.03] px-5 py-2.5 text-base"
-          />
-        </FormControl>
-        {description && <FormDescription>{description}</FormDescription>}
-        <FormMessage />
-      </FormItem>
-    )}
-  />
-);
