+++ src/components/project/case-study.tsx
@@ -1,6 +1,13 @@
 import { motion, animate, easeOut, useInView } from "framer-motion";
 import { CaseStudy, Clients } from "@/sanity/types";
-import { useCallback, useEffect, useState, useRef, Fragment } from "react";
+import {
+  useCallback,
+  useEffect,
+  useState,
+  useRef,
+  Fragment,
+  forwardRef,
+} from "react";
 import { AnimatePresence } from "framer-motion";
 import { useRouter } from "next/router";
 import { MediaRenderer } from "@/components/media-renderer";
@@ -103,143 +110,154 @@ function MediaGroup({
 
 // Navigation Component - Provides UI for navigating between media groups
 // Handles expanded/collapsed states and animations
-const Navigation = ({
-  activeGroup,
-  groups,
-  isExpanded,
-  onToggleExpand,
-  onNext,
-  onPrev,
-  canGoNext,
-  canGoPrev,
-}: NavigationProps) => {
-  const [prevIndex, setPrevIndex] = useState(activeGroup);
-  const direction = activeGroup > prevIndex ? 1 : -1;
-  const currentGroup = groups[activeGroup];
+const Navigation = forwardRef<HTMLDivElement, NavigationProps>(
+  (
+    {
+      activeGroup,
+      groups,
+      isExpanded,
+      onToggleExpand,
+      onNext,
+      onPrev,
+      canGoNext,
+      canGoPrev,
+    },
+    ref,
+  ) => {
+    const [prevIndex, setPrevIndex] = useState(activeGroup);
+    const direction = activeGroup > prevIndex ? 1 : -1;
+    const currentGroup = groups[activeGroup];
 
-  useEffect(() => {
-    setPrevIndex(activeGroup);
-  }, [activeGroup]);
+    useEffect(() => {
+      setPrevIndex(activeGroup);
+    }, [activeGroup]);
 
-  const transition = {
-    duration: 0.4,
-    ease: EASINGS.easeOutQuart,
-  };
+    const transition = {
+      duration: 0.4,
+      ease: EASINGS.easeOutQuart,
+    };
 
-  const originStyles = { transformOrigin: "bottom center" };
+    const originStyles = { transformOrigin: "bottom center" };
 
-  return (
-    <motion.div
-      layout
-      layoutId="nav-root"
-      className="fixed bottom-0 left-0 right-0 z-50 px-3 pb-[calc(0.5rem_+_env(safe-area-inset-bottom))] md:bottom-4 md:px-4"
-      style={originStyles}
-      initial={{ opacity: 0, y: 20, scale: 0.9 }}
-      animate={{ opacity: 1, y: 0, scale: 1 }}
-      exit={{ opacity: 0, y: 20, scale: 0.9 }}
-      transition={transition}
-    >
+    return (
       <motion.div
+        ref={ref}
         layout
-        layoutId="nav-container"
-        className="mx-auto h-min w-min overflow-hidden bg-white/60 p-1 drop-shadow-2xl backdrop-blur-2xl"
-        animate={{
-          borderRadius: isExpanded ? 32 : 16,
-        }}
-        transition={transition}
+        layoutId="nav-root"
+        className="fixed bottom-0 left-0 right-0 z-50 px-3 pb-[calc(0.5rem_+_env(safe-area-inset-bottom))] md:bottom-4 md:px-4"
         style={originStyles}
+        initial={{ opacity: 0, y: 20, scale: 0.9 }}
+        animate={{ opacity: 1, y: 0, scale: 1 }}
+        exit={{ opacity: 0, y: 20, scale: 0.9 }}
+        transition={transition}
       >
         <motion.div
           layout
-          layoutId="nav-background"
-          className="relative overflow-hidden bg-white"
+          layoutId="nav-container"
+          className="mx-auto h-min w-min overflow-hidden bg-white/60 p-1 drop-shadow-2xl backdrop-blur-2xl"
           animate={{
-            borderRadius: isExpanded ? 28 : 12,
+            borderRadius: isExpanded ? 32 : 16,
           }}
           transition={transition}
           style={originStyles}
         >
-          <AnimatePresence mode="popLayout" initial={false}>
-            {!isExpanded ? (
-              <motion.button
-                layout
-                key="collapsed"
-                className="flex w-full items-center justify-between gap-2 px-4 py-2 focus:outline-none"
-                onClick={onToggleExpand}
-                initial={{ opacity: 0 }}
-                animate={{ opacity: 1 }}
-                exit={{ opacity: 0 }}
-                transition={transition}
-              >
-                <span className="opacity-50">{currentGroup.heading}</span>
-                <XIcon className="h-4 w-4 rotate-45 opacity-60" />
-              </motion.button>
-            ) : (
-              <motion.div
-                layout
-                key="expanded"
-                className="relative w-screen max-w-[calc(100vw-2rem)] md:max-w-[400px]"
-                initial={{ opacity: 0 }}
-                animate={{ opacity: 1 }}
-                exit={{ opacity: 0 }}
-                transition={transition}
-              >
-                <div className="flex w-full flex-col gap-3 px-6 py-6">
-                  <AnimatePresence mode="popLayout" custom={direction}>
-                    <motion.div
-                      key={`content-${activeGroup}`}
-                      initial={{ opacity: 0, y: direction * 20 }}
-                      animate={{ opacity: 1, y: 0 }}
-                      exit={{ opacity: 0, y: direction * -20 }}
-                      transition={transition}
-                    >
-                      {currentGroup.heading && (
-                        <div className="mb-2 flex w-full items-center justify-between text-lg">
-                          <h4 className="opacity-70">{currentGroup.heading}</h4>
-                        </div>
-                      )}
-                      {currentGroup.caption && (
-                        <p className="text-sm opacity-50">
-                          {currentGroup.caption}
-                        </p>
-                      )}
-                    </motion.div>
-                  </AnimatePresence>
-
-                  <div className="flex gap-1">
-                    <button
-                      onClick={onPrev}
-                      disabled={!canGoPrev}
-                      className="rounded-full bg-white/80 p-2 transition-opacity hover:bg-white/90 disabled:opacity-30"
-                    >
-                      <ChevronUpIcon className="h-4 w-4" />
-                    </button>
-                    <button
-                      onClick={onNext}
-                      disabled={!canGoNext}
-                      className="rounded-full bg-white/80 p-2 transition-opacity hover:bg-white/90 disabled:opacity-30"
-                    >
-                      <ChevronDownIcon className="h-4 w-4" />
-                    </button>
-                  </div>
-                </div>
-
-                <button
+          <motion.div
+            layout
+            layoutId="nav-background"
+            className="relative overflow-hidden bg-white"
+            animate={{
+              borderRadius: isExpanded ? 28 : 12,
+            }}
+            transition={transition}
+            style={originStyles}
+          >
+            <AnimatePresence mode="popLayout" initial={false}>
+              {!isExpanded ? (
+                <motion.button
+                  layout
+                  key="collapsed"
+                  className="flex w-full items-center justify-between gap-2 px-4 py-2 focus:outline-none"
                   onClick={onToggleExpand}
-                  className={`absolute right-6 ${
-                    currentGroup.heading ? "top-[1.75rem]" : "top-[2.25rem]"
-                  }`}
+                  initial={{ opacity: 0 }}
+                  animate={{ opacity: 1 }}
+                  exit={{ opacity: 0 }}
+                  transition={transition}
                 >
-                  <XIcon className="h-4 w-4 opacity-60" />
-                </button>
-              </motion.div>
-            )}
-          </AnimatePresence>
+                  <span className="opacity-50">{currentGroup.heading}</span>
+                  <XIcon className="h-4 w-4 rotate-45 opacity-60" />
+                </motion.button>
+              ) : (
+                <motion.div
+                  layout
+                  key="expanded"
+                  className="relative w-screen max-w-[calc(100vw-2rem)] md:max-w-[400px]"
+                  initial={{ opacity: 0 }}
+                  animate={{ opacity: 1 }}
+                  exit={{ opacity: 0 }}
+                  transition={transition}
+                >
+                  <div className="flex w-full flex-col gap-3 px-6 py-6">
+                    <AnimatePresence mode="popLayout" custom={direction}>
+                      <motion.div
+                        key={`content-${activeGroup}`}
+                        initial={{ opacity: 0, y: direction * 20 }}
+                        animate={{ opacity: 1, y: 0 }}
+                        exit={{ opacity: 0, y: direction * -20 }}
+                        transition={transition}
+                      >
+                        {currentGroup.heading && (
+                          <div className="mb-2 flex w-full items-center justify-between text-lg">
+                            <h4 className="opacity-70">
+                              {currentGroup.heading}
+                            </h4>
+                          </div>
+                        )}
+                        {currentGroup.caption && (
+                          <p className="text-sm opacity-50">
+                            {currentGroup.caption}
+                          </p>
+                        )}
+                      </motion.div>
+                    </AnimatePresence>
+
+                    <div className="flex gap-1">
+                      <button
+                        onClick={onPrev}
+                        disabled={!canGoPrev}
+                        className="rounded-full bg-white/80 p-2 transition-opacity hover:bg-white/90 disabled:opacity-30"
+                      >
+                        <ChevronUpIcon className="h-4 w-4" />
+                      </button>
+                      <button
+                        onClick={onNext}
+                        disabled={!canGoNext}
+                        className="rounded-full bg-white/80 p-2 transition-opacity hover:bg-white/90 disabled:opacity-30"
+                      >
+                        <ChevronDownIcon className="h-4 w-4" />
+                      </button>
+                    </div>
+                  </div>
+
+                  <button
+                    onClick={onToggleExpand}
+                    className={`absolute right-6 ${
+                      currentGroup.heading ? "top-[1.75rem]" : "top-[2.25rem]"
+                    }`}
+                  >
+                    <XIcon className="h-4 w-4 opacity-60" />
+                  </button>
+                </motion.div>
+              )}
+            </AnimatePresence>
+          </motion.div>
         </motion.div>
       </motion.div>
-    </motion.div>
-  );
-};
+    );
+  },
+);
+
+// Add display name for better debugging
+Navigation.displayName = "Navigation";
 
 // Main Component - Orchestrates the entire case study experience
 export default function ProjectCaseStudy({
